Revert relocation of transform utility functions
