+++ cstar/tests/unit_tests/orchestration/test_orchestration.py
@@ -9,10 +9,7 @@ import networkx as nx
 import pytest
 
 from cstar.applications.roms_marbl.models import RomsMarblBlueprint
-from cstar.applications.roms_marbl.transforms import (
-    RomsMarblTimeSplitter,
-    get_time_slices,
-)
+from cstar.applications.roms_marbl.transforms import RomsMarblTimeSplitter
 from cstar.base.env import ENV_CSTAR_RUNID, FLAG_ON
 from cstar.base.feature import ENV_FF_ORCH_TRX_TIMESPLIT
 from cstar.orchestration.launch.local import LocalLauncher
@@ -26,9 +23,7 @@ from cstar.orchestration.orchestration import (
     Status,
 )
 from cstar.orchestration.serialization import deserialize
-from cstar.orchestration.transforms import (
-    WorkplanTransformer,
-)
+from cstar.orchestration.transforms import WorkplanTransformer, get_time_slices
 
 if t.TYPE_CHECKING:
     from collections.abc import Iterable
