+++ cstar/tests/unit_tests/orchestration/test_splitting.py
@@ -6,15 +6,13 @@ from unittest import mock
 
 import pytest
 
-from cstar.applications.roms_marbl.transforms import (
-    RomsMarblTimeSplitter,
-    get_time_slices,
-)
+from cstar.applications.roms_marbl.transforms import RomsMarblTimeSplitter
 from cstar.base.env import ENV_CSTAR_DATA_HOME, ENV_CSTAR_RUNID
 from cstar.base.utils import DEFAULT_OUTPUT_ROOT_NAME
 from cstar.orchestration.models import Application, Step, Workplan
 from cstar.orchestration.orchestration import LiveStep
 from cstar.orchestration.transforms import (
+    get_time_slices,
     get_transforms,
 )
